Delete profesor arreglado

+++ src/main/java/com/example/demo/services/UsuarioService.java
@@ -7,7 +7,7 @@ import com.example.demo.entities.User;
 import com.example.demo.models.UserModel;
 public interface UsuarioService {
 	public abstract List<UserModel> listAllProfesores();
-	public abstract int removeProfesor(int id);
+	public abstract int removeProfesor(long id);
 	public abstract User transform(UserModel userModel);
 	public abstract UserModel transform(User user);
